refactor(plugin-image-gallery): Combine helper files, cleanup
